requirements.txt
flask>=3.0.0
python-dotenv>=1.0.0
anthropic>=0.40.0
tqdm>=4.66.0

# Whisper — install separately (already done):
# cd "C:\המחשב שלי\Desktop\whisper" && pip install -e .
